requirements.txt
matplotlib
tensorflow
tensorboard
numpy
inflect
librosa
scipy
pillow
PyYAML
tqdm
torch
torchaudio
torchvision
num2words


# My libraries with their versions
# Python 3.11.9 Windows Driver Version: 566.36 CUDA Version: 12.7

#inflect==7.5.0
#librosa==0.11.0
#matplotlib==3.10.1
#numpy==2.1.3
#pillow==11.2.1
#PyYAML==6.0.2
#scipy==1.15.2
#tensorboard==2.19.0
#tensorflow==2.19.0
#torch==2.4.1+cu124
#torchaudio==2.4.1+cu124
#torchvision==0.19.1+cu124
#tqdm==4.67.1
